refactor: consolidate manifest resolution and migrate console.* to logger
Extract shared manifest resolution types (ContentRef, OutputManifest) and
functions (isManifestHash, resolveContentRef, resolveDataBundle,
resolveManifest) into lib/manifest-resolution.ts, eliminating ~150 LOC of
duplication between materialize-cells.ts and useManifestResolver.ts.

This also removes an unsafe `as Record<string, unknown>` type cast that
was used to set display_id in useManifestResolver.ts — the shared
resolveManifest now handles display_id directly via the JupyterOutput type
which already supports the field.

Migrate all raw console.* calls to the logger abstraction across 7 files:
useCondaDependencies, useDependencies, useDenoDependencies, useGitInfo,
useTrust, useUpdater, MarkdownCell. The logger gates debug output in
production builds.
